+++ sdk/src/main/java/com/uber/sdk/rides/client/SessionConfiguration.java
@@ -31,7 +31,7 @@ import java.util.Locale;
 
 import javax.annotation.Nonnull;
 
-import static com.uber.sdk.rides.client.SessionConfiguration.EndpointRegion.WORLD;
+import static com.uber.sdk.rides.client.SessionConfiguration.EndpointRegion.DEFAULT;
 import static com.uber.sdk.rides.client.utils.Preconditions.checkNotNull;
 
 /**
@@ -55,13 +55,19 @@ public class SessionConfiguration implements Serializable {
     }
 
     public enum EndpointRegion {
-        WORLD("uber.com"),
-        CHINA("uber.com.cn");
+        DEFAULT("uber.com");
 
-        public String domain;
+        private String domain;
 
         EndpointRegion(String domain) {
             this.domain = domain;
+        }
+
+        /**
+         * @return domain to use.
+         */
+        public String getDomain() {
+            return domain;
         }
     }
 
@@ -73,7 +79,6 @@ public class SessionConfiguration implements Serializable {
         private String clientSecret;
         private String serverToken;
         private String redirectUri;
-        private EndpointRegion region = EndpointRegion.WORLD;
         private Environment environment;
         private Collection<Scope> scopes;
         private Collection<String> customScopes;
@@ -127,17 +132,6 @@ public class SessionConfiguration implements Serializable {
         }
 
         /**
-         * Set the {@link EndpointRegion} your app is registered in.
-         * Used to determine what endpoints to send requests to.
-         *
-         * @param region The {@link EndpointRegion} the SDK should use
-         */
-        public Builder setEndpointRegion(@Nonnull EndpointRegion region) {
-            this.region = region;
-            return this;
-        }
-
-        /**
          * Sets the {@link Environment} to be used for API requests
          *
          * @param environment to be set
@@ -187,10 +181,6 @@ public class SessionConfiguration implements Serializable {
         public SessionConfiguration build() {
             checkNotNull(clientId, "Client must be set");
 
-            if (region == null) {
-                region = WORLD;
-            }
-
             if (environment == null) {
                 environment = Environment.PRODUCTION;
             }
@@ -216,7 +206,7 @@ public class SessionConfiguration implements Serializable {
                     clientSecret,
                     serverToken,
                     redirectUri,
-                    region,
+                    DEFAULT,
                     environment,
                     scopes,
                     customScopes,
@@ -228,7 +218,7 @@ public class SessionConfiguration implements Serializable {
     private final String clientSecret;
     private final String serverToken;
     private final String redirectUri;
-    private final EndpointRegion region;
+    private final EndpointRegion endpointRegion;
     private final Environment environment;
     private final Collection<Scope> scopes;
     private final Collection<String> customScopes;
@@ -238,7 +228,7 @@ public class SessionConfiguration implements Serializable {
                                    @Nonnull String clientSecret,
                                    @Nonnull String serverToken,
                                    @Nonnull String redirectUri,
-                                   @Nonnull EndpointRegion region,
+                                   @Nonnull EndpointRegion endpointRegion,
                                    @Nonnull Environment environment,
                                    @Nonnull Collection<Scope> scopes,
                                    @Nonnull Collection<String> customScopes,
@@ -247,7 +237,7 @@ public class SessionConfiguration implements Serializable {
         this.clientSecret = clientSecret;
         this.serverToken = serverToken;
         this.redirectUri = redirectUri;
-        this.region = region;
+        this.endpointRegion = endpointRegion;
         this.environment = environment;
         this.scopes = scopes;
         this.customScopes = customScopes;
@@ -291,16 +281,6 @@ public class SessionConfiguration implements Serializable {
     }
 
     /**
-     * Gets the current {@link EndpointRegion} the SDK is using.
-     * Defaults to {@link EndpointRegion#WORLD}.
-     *
-     * @return The {@link EndpointRegion} the SDK is using.
-     */
-    public EndpointRegion getEndpointRegion() {
-        return region;
-    }
-
-    /**
      * Gets the environment configured, either {@link Environment#PRODUCTION} or {@link Environment#SANDBOX}
      *
      * @return {@link Environment} that is configured
@@ -310,11 +290,29 @@ public class SessionConfiguration implements Serializable {
     }
 
     /**
+     * Gets the current {@link EndpointRegion} the SDK is using.
+     * Defaults to {@link EndpointRegion#DEFAULT}.
+     *
+     * @return the {@link EndpointRegion} the SDK us using.
+     */
+    public EndpointRegion getEndpointRegion() {
+        return endpointRegion;
+    }
+
+    /**
      * Gets the endpoint host used to hit the Uber API.
      */
     @Nonnull
     public String getEndpointHost() {
-        return String.format("https://%s.%s", environment.subDomain, region.domain);
+        return String.format("https://%s.%s", environment.subDomain, DEFAULT.getDomain());
+    }
+
+    /**
+     * Gets the login host used to sign in to the Uber API.
+     */
+    @Nonnull
+    public String getLoginHost() {
+        return String.format("https://login.%s", DEFAULT.getDomain());
     }
 
     /**
@@ -347,7 +345,6 @@ public class SessionConfiguration implements Serializable {
         return new Builder()
                 .setClientId(clientId)
                 .setRedirectUri(redirectUri)
-                .setEndpointRegion(region)
                 .setEnvironment(environment)
                 .setScopes(scopes);
     }